# improve speed of RuboCop::PathUtil#smart_path