@Drenmi 2017-11-24
Nice! 👍

I think rebasing on latest master should fix the CI failure.