@pocke 2017-11-25
Awesome improvement!

Fix a rubocop's offense to fix the build failure, and resolve the conflict.
Thank you!