@walf443 2017-11-25
@pocke  sorry. fixed it.

travis-ci/pr seems to failed followings:

```
$ bundle exec rake documentation_syntax_check generate_cops_documentation
Files:         458
Modules:        76 (   12 undocumented)
Classes:       430 (    2 undocumented)
Constants:     666 (  661 undocumented)
Attributes:     23 (    0 undocumented)
Methods:      1019 (  850 undocumented)
 31.12% documented
* generated /home/travis/build/bbatsov/rubocop/manual/cops_bundler.md
* generated /home/travis/build/bbatsov/rubocop/manual/cops_gemspec.md
* generated /home/travis/build/bbatsov/rubocop/manual/cops_layout.md
* generated /home/travis/build/bbatsov/rubocop/manual/cops_lint.md
* generated /home/travis/build/bbatsov/rubocop/manual/cops_metrics.md
* generated /home/travis/build/bbatsov/rubocop/manual/cops_naming.md
* generated /home/travis/build/bbatsov/rubocop/manual/cops_performance.md
* generated /home/travis/build/bbatsov/rubocop/manual/cops_rails.md
* generated /home/travis/build/bbatsov/rubocop/manual/cops_security.md
* generated /home/travis/build/bbatsov/rubocop/manual/cops_style.md
git diff --quiet manual
git diff manual
diff --git a/manual/cops_rails.md b/manual/cops_rails.md
index 0c7a4af..4f076b6 100644
--- a/manual/cops_rails.md
+++ b/manual/cops_rails.md
@@ -203,11 +203,11 @@ create_table :users do |t|
 end
 
 
-### Important attributes
+### Configurable attributes
 
-Attribute | Value
---- | ---
-Include | db/migrate/\*.rb
+Name | Default value | Configurable values
+--- | --- | ---
+Include | `db/migrate/*.rb` | Array
```

Is it related this PR?