@walf443 2017-11-25
I send PR to fix this https://github.com/bbatsov/rubocop/pull/5124/files